+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- Various issues with repeat + button in Pages Mode when repeat=page.
+
 [4.41.2] - 2017-11-08
 ---------------------
 ##### Changed